refactored teacher and student creation in signup
